media: gsc-out: add current index check routine to gscaler's ISR
Gscaler only uses active_buf_q to check if there is a queued
job. But it's possible for gscaler interrupt to be asserted while
adding vb2_buffer to active_buf_q. If so, gscaler isr would be
executed right after that. But driver checks if it's okay to pass
done buffer to user. It caused page fault. Because user can queue
buffer that is running.
This patch added check routine whether current running buffer is
that same with done buffer.

Change-Id: I33f574867efdce197433bb9285cde1353dfbba57
Signed-off-by: Hyunwoong Kim <khw0178.kim@samsung.com>