crypto: ACE: modify chained scatterlist handling
This patch modifies the way to handle chained scatterlist,
because this driver handled it in a differnet way from kernel crypto API

eseqiv_givencrypt() calls ACE driver when VPN use the driver.
eseqiv_givencrypt() gives the data to the driver in the type of chained scatterlist.
In eseqiv_givencrypt(), chained scatterlist is marked by setting the length 0.
And, it uses scatterwalk_sg_next() to go to the next entry.
In scatterwalk_sg_next(), if the length is 0,
then it's handled as chained scatterlist.

However, ACE driver used sg_next() to go to next entry.
In sg_next(), last two bit of page link is cheked.
If it is 0x01, the value that indicates chained scatterlist,
it's handled as chained scatterlist.

Change-Id: Ic2cd79c8cde9d7a59197ea6973f3b23c6e59d245
Signed-off-by: Jihyun Kim <jihyun86.kim@samsung.com>