crypto: ACE: fix a bug making system hang
This patch fixes system hang.
System hang could be happened when user called get_sync_lock, 10
because there was the moment that FLAGS_BC_BUSY was set in get_sync_lock()
even if H/W was not being used.
If FLAGS_BC_BUSY is set, request is not proceed but enqueued.
And, the request is not proceed till next request is entered.
If next reqeust is not entered, user kept waiting the complete signal and occured system hang.

Change-Id: I089a4b9891b9ab36adac971dad378f437afed9d4
Signed-off-by: Jihyun Kim <jihyun86.kim@samsung.com>