usb: GADGET: remove bMaxpower in Android Device descriptor
This patch remove bMaxPower in Android Device descriptor.
In USB 2.0 spec, MaxPower value is same as bMaxPower * 2 mA.
But, in USB 3.0 spec, MaxPower value is same as bMaxPower * 8 mA.
The bMaxPower of Android Device descriptor is set as 0xFA.
If MaxPower is bigger than 900mA in Superspeed, USB is only operated in Highspeed.

Change-Id: Ic442b19020b9104345f2604fc0c5796510730e03
Signed-off-by: Seokbae Yoon <seokbae.yoon@samsung.com>