MFC: MEDIA: EXYNOS: Remove unnessecary condition check
This patch removes unnessecary condition check routine
in s_ctrl() of the decoder. The validataion of a control
value be verified already. And we don't need to check
codec-specific control because it's meaningful only
when a selected codec is matched, also the check can
make a limitation of the decoding sequence.

Change-Id: Ib95eed35083a3cf16449081aaebe1cda95efa739
Signed-off-by: Jeongtae Park <jtp.park@samsung.com>