usb: s3c_udc: Change from legacy cache api to non-cache allocation
This patch remove legacy coherence function for dma operation.
This functions are private support functions
These are not part of the API and are liable to change.
Drivers must not use these.
These are used to send ep0 data packet.
GET_STATUS and SET_TEST_PACKET requests are handled in udc.
But, dma map/unmap is used for gadget request, not udc.
So, buf is allocated by dma_alloc_coherent() for ep0 data phase.

Change-Id: I76c2a47f3be55a7744a23f22cc9b78ae6358ad32
Signed-off-by: Yulgon Kim <yulgon.kim@samsung.com>