ARM: SAMSUNG: fix probe sequence for ADC driver
The request_irq function was called earlier than ioremap function
in the probe function of the old ADC driver. It was possible to
occur a kernel panic because the adc->regs value was used by
s3c_adc_irq function without ioremap. This patch changes the probe
sequence to call ioremap first.

Change-Id: Ia37452117f5858131c74ada4fc28723f755079f7
Signed-off-by: Eunki Kim <eunki_kim@samsung.com>