USB: DWC3: Role Switch driver for Exynos USB3 DRD
This is role switch driver for the Exynos USB3 DRD core.
The driver implements the otg final state machine and control the
activation of the device controller or host controller.

The driver determines the DRD role according to the ID pin as reflected
by ID gpio state. The VBus GPIO is used to detect valid B-Session in
B-Device mode.

Code related to DRD switch functionality was removed from Exynos
USB Switch driver.

Change-Id: I7cd54f7687ef2c7f745d51bc7cf003224f64b928
Signed-off-by: Seokbae Yoon <seokbae.yoon@samsung.com>