ARM: EXYNOS: update LPA feature
Update LPA feature for fully support to low power idle feature.
But currently, the LPA mode is not operate as disable feature.
Because need to check the operation of USB, GPS and BT for enter
LPA mode. So, these drivers should be update operation status to
the cpuidle driver. This feature will enable after these driver's
update completed.

Change-Id: Ie91b4d22cc734b17c7be49aee81c019ccd15d8f5
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>