mmc: dw_mmc: read all data in FIFO after Data transfer over interrupt in pio mode
In dwc manual, the below contents are described.
"During end of packet, interrupt is not generated if threshold
programming is larger than any remaining data. It is responsibility
of host to read remaining bytes on seeing Data Transfer Done
interrupt"
We also have seen the data cannot be read fully
when "sg_miter->length" is less than FIFO size.

Change-Id: Ie089b08974d6d66c24bc7a7781fcc209217e2335
Signed-off-by: Kyoungil Kim <ki0351.kim@samsung.com>