USB: exynos_ss_udc: Fix EP0 stall handling
DWC3 databook says "When device software sets STALL with the Set Stall
command, it is possible that this command can be delayed on the SoC bus
due to latency or the SoC bus servicing other agents. In this case,
software can receive an XferNotReady event if the core detects no TRB
active before it detected the Set Stall." In practice this can lead to
multiple EP0 restarts on receiving XferNotReady event when current
driver's EP0 state doesn't match to stage requested by host. This patch
ignores such XferNotReady events and waits until "the core clears the
STALL automatically when it receives a SETUP token for the endpoint".
If Set Stall was issued by gadget layer, Setup transfer must be started
on XferNotReady event to receive future Setup request from host.

Change-Id: I4fa838bd6388f638f63666f2395a670e85003e16
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>