USB: exynos-drd: Apply the w/a for HOST ACK TP Retry
DWC3 revisions from 1.90a to 2.10a have a bug.
For ss bulk-in data packet, when the host detects
a DPP error or the internal buffer becomes full,
it retries with an ACK TP Retry=1. However, under the following
conditions, the Retry=1 is falsely carried over to the next burst.
If there is only single active asynchronous SS EP at the time,
the active asynchronous EP is a Bulk IN EP and
the burst with the correctly Retry=1 ACK TP and the next burst
belong to the same transfer.

This patch is the workaround for this bug.

Change-Id: Ia55deb9a01a96ef1db5246ed1ebb674359a7edf1
Signed-off-by: HoJune Byun <hjune.byun@samsung.com>
Signed-off-by: Yulgon Kim <yulgon.kim@samsung.com>