usb: xhci-exynos: Fix the bug in PORTSC.PLS
When the port power is switched on to a 2.0 port and the device
is not connected, the PLS field in PORTSC does not get updated
from 4'h4 to 4'h5. It affects the version 2.00a and earlier.
The workaround is as follows;
Before switching on the port power, set the GUSB2PHYCFG[6] to 0.
After switching on the port power, set this bit to 1.
This patch is the workaround for this bug.

Change-Id: I1c0ab6c1e117a3cb1a4caf2dabb8c36d1e20505d
Signed-off-by: HoJune Byun <hjune.byun@samsung.com>
Signed-off-by: Yulgon Kim <yulgon.kim@samsung.com>