USB: exynos_ss_udc: Fix pullup method
According to DWC3 databook, if the application wants to disconnect
from the host, it should end all active transfers and clear RunStop
bit. If the application decides it wants to reconnect to the host,
it must follow the steps as on Device Power-On or Soft Reset.
Additional change relates to UDC stop method. The controller
requires event count to be 0 otherwise DSTS.DevCtrlHlt will never
be set.

Change-Id: I0c8153e6135d9246c0630aac808d93203ede3619
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>