USB: exynos_ss_udc: Mask events and flush FIFOs when disable udc
On cable connection DRD core will resume first and enable
IRQ, while UDC is not initialized yet and its state might be
inconsistent/corrupted. On practice this results in
controller(?) malfunction, when interrupts are continuously
generated, and as a consiquence device freeze. This patch
disables events and flushes FIFOs during disabling udc to
prevent this situation.

Change-Id: I027f5380e282b4670db8cc6a017005f3f1b24f9f
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>