USB: exynos-drd: Use bus clock for RAM instead of pipe clock
This patch fixes the issue when reading XHCI registers after system
resume returns all zeros. In Host mode, the ram_clk must be assigned
to the bus_clk only. The reason is if the SS port 0 goes to P3, the
pipe_clk will be shut down, and USB2.0 ports will not be able to
operate.

Change-Id: I3465f201f16fc9d681c288637f86f954a4eecaac
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>